THRIFT-916 Wall_Wextra_pedantic_Wno-long-long_Wno-variadic-macros_Wno-overflow_NOWARNINGS_DebianLenny.patch applied
git-svn-id: https://svn.apache.org/repos/asf/thrift/trunk@1037127 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/cpp/src/transport/THttpTransport.cpp
@@ -127,9 +127,9 @@ uint32_t THttpTransport::parseChunkSize(char* line) {
   if (semi != NULL) {
     *semi = '\0';
   }
-  int size = 0;
+  uint32_t size = 0;
   sscanf(line, "%x", &size);
-  return (uint32_t)size;
+  return size;
 }
 
 uint32_t THttpTransport::readContent(uint32_t size) {
